added init, renamed ZMQ server
